don't use docker-entrypoint.sh from base C* image
The entrypoint script from the base image causing a problem with configuring
seeds. It sets seeds in cassandra.yaml to IP address used for the
listen_address property. I want it set to a k8s service though. I decided to
simply not call the entrypoint script and things are working (so far) better
than expected.